test: share async signal timeout helper
